src/test/java/lotto/domain/RankTest.java
package lotto.domain;

import org.junit.jupiter.api.Test;

import static org.assertj.core.api.Assertions.assertThat;

class RankTest {
    @Test
    void RANK_FIRST_TEST() {
        assertThat(Rank.valueOf(6, false)).isEqualTo(Rank.FIRST);
    }

    @Test
    void RANK_SECOND_TEST() {
        assertThat(Rank.valueOf(5, true)).isEqualTo(Rank.SECOND);
    }

    @Test
    void RANK_THIRD_TEST() {
        assertThat(Rank.valueOf(5, false)).isEqualTo(Rank.THIRD);
    }

    @Test
    void RANK_FOURTH_TEST() {
        assertThat(Rank.valueOf(4, false)).isEqualTo(Rank.FOURTH);
    }

    @Test
    void RANK_FIFTH_TEST() {
        assertThat(Rank.valueOf(3, false)).isEqualTo(Rank.FIFTH);
    }
}